Introduce git-sorthist script
This change introduces the git-sorthist script. It can be used to
reorder the commit dates (not the commits themselves) to be
chronologically correct.
When developing a feature it can take some time until the final set of
commits emerges. Before that, changes may get reordered more or less
arbitrarily. There is some value in having a clean git history and it
can be argued that the commits in such a history should be ordered
chronologically. This is where this script comes in: it allows the user
to sort the commit dates accordingly.