Introduce git-redatehist script
This change introduces another script to rewrite git history:
git-redatehist. It can be used to adjust commit and author dates for a
sequence of commits. Changes will be made in such a way that starting
from a provided date, each commit will use the next date. The times will
be random time (within constraints). Time will only be changed if the
commit's date is not already as expected.