Adjust git-sorthist to set all relevant dates
In most if not all cases, when we sort the commits chronologically using
git-sorthist, we don't just want one of the dates to be set. Rather,
author, committer, and signature date are likely intended to be set in
unison. Right now that can already be accomplished with another rebase
operation following a git-sorthist invocation, but that takes a while
and requires the new commit SHA-1 to be looked up first.
To omit this extra work, this change adjusts the script to set all three
dates rather than just the author date.